backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-rest-framework>=0.1.0",
    "gunicorn>=23.0.0",
    "markdown>=3.9",
    "questionary>=2.0.0",
    "whitenoise>=6.11.0",
]

[tool.setuptools.packages.find]
include = ["api", "configuration"]
